refactore key_type placement
