+++ core/crypto/crypto_store.hpp
@@ -12,8 +12,8 @@
 #include <boost/optional.hpp>
 #include <libp2p/crypto/key.hpp>
 #include "crypto/bip39/bip39_provider.hpp"
+#include "crypto/crypto_store/key_type.hpp"
 #include "crypto/ed25519_types.hpp"
-#include "crypto/key_type.hpp"
 #include "crypto/secp256k1_types.hpp"
 #include "crypto/sr25519_types.hpp"
