+++ test/core/crypto/CMakeLists.txt
@@ -5,7 +5,7 @@
 
 add_subdirectory(bip39)
 add_subdirectory(blake2)
-add_subdirectory(crypto_storage_key_type)
+add_subdirectory(crypto_store)
 add_subdirectory(ed25519)
 add_subdirectory(keccak)
 add_subdirectory(hasher)
