+++ test/core/crypto/crypto_store/CMakeLists.txt
@@ -4,8 +4,8 @@
 #
 
 addtest(key_type_test
-        key_type_test.cpp
-        )
+    key_type_test.cpp
+    )
 target_link_libraries(key_type_test
-        crypto_store_key_type
-        )
+    crypto_store_key_type
+    )
